+++ MathExpressionEvaluator/Expressions/Unary/Concrete/LnUnaryExpression.cs
@@ -9,9 +9,9 @@ namespace MathExpressionEvaluator.Expressions.Unary.Concrete
         {
         }
 
-        internal override decimal Evaluate()
+        internal override double Evaluate()
         {
-            return (decimal) Math.Log((double) Expression.Evaluate(), Math.E);
+            return Math.Log(Expression.Evaluate(), Math.E);
         }
     }
 }
