+++ MathExpressionEvaluator/Expressions/Value/NaturalExpression.cs
@@ -4,9 +4,9 @@ namespace MathExpressionEvaluator.Expressions.Value
 {
     internal sealed class NaturalExpression : Expression
     {
-        internal override decimal Evaluate()
+        internal override double Evaluate()
         {
-            return (decimal) Math.E;
+            return Math.E;
         }
     }
 }
